# api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
tags:
- name: Actualizar datos de usuario
  description: Actualiza los datos de la cuenta
- name: Recuperacion de clave
  description: Hace que se envie un codigo de verificacion al correo de la cuenta
- name: Login de usuario
  description: Permite iniciar sesion al usario
- name: Cambio de clave
  description: Cambia la clave un usuario
- name: Obtener Usuarios
  description: Obtiene parte de los usuarios del sistema
- name: Registro de usuarios
  description: Registra un nuevo usuario
paths:
  /usuario/actualizar:
    put:
      tags:
      - Actualizar datos de usuario
      summary: ' Actualizar usuario'
      description: Actualiza los datos de un usuario existente
      operationId: actualizarDatos
      parameters:
      - in: query
        description: Nuevos datos a actualizar
        required: true
        example: "usuario:'Juan',correo:'juanmanuel200413@gmail.com',clave:'1234'"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DatosUsuario"
        required: true
      responses:
        "400":
          description: "Atributos de usuario, correo contraseña son obligatorios"
          content:
            '*/*':
              schema:
                type: object
        "200":
          description: Usuario actualizado
          content:
            application/json:
              schema:
                type: string
        "404":
          description: Usuario no encontrado
          content:
            '*/*':
              schema:
                type: object
      security:
      - bearerAuth: []
  /publico/registro:
    post:
      tags:
      - Registro de usuarios
      summary: Registrar usuario
      description: Registra un nuevo usuario en la base de datos
      operationId: registrarUsuario
      parameters:
      - in: query
        description: Datos del usuario a registrar
        required: true
        example: "usuario:'Juan',correo:'juanmanuel200413@gmail.com',clave:'1234'"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DatosUsuario"
        required: true
      responses:
        "400":
          description: "Atributos de usuario, correo contraseña son obligatorios"
          content:
            '*/*':
              schema:
                type: object
        "200":
          description: Usuario registrado
          content:
            application/json:
              schema:
                type: string
        "409":
          description: El usuario ya existe
          content:
            '*/*':
              schema:
                type: object
  /publico/recuperarClave:
    post:
      tags:
      - Recuperacion de clave
      summary: Recuperar clave
      description: Inicia el proceso para poder obtner una nueva cuenta del usuario
      operationId: recuperarClave
      parameters:
      - in: query
        description: Nombre del usaurio que se quiere recuperar clave
        required: true
        example: Juan
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: Codigo generado
          content:
            application/json:
              schema:
                type: string
        "404":
          description: Usuario no existente
          content:
            '*/*':
              schema:
                type: object
        "500":
          description: Error al recuperar el usuario
          content:
            '*/*':
              schema:
                type: object
  /publico/login:
    post:
      tags:
      - Login de usuario
      summary: Iniciar sesion
      description: Genera el token de autenticacion para el usuario
      operationId: login
      parameters:
      - in: query
        description: Datos del usuario que va realizar el inicio de sesion
        required: true
        example: "usuario:'Juan',correo:'juanmanuel200413@gmail.com',clave:'1234'"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DatosUsuario"
        required: true
      responses:
        "404":
          description: No existe el usuario con los datos indicados
          content:
            '*/*':
              schema:
                type: object
        "200":
          description: Token generado
          content:
            application/json:
              schema:
                type: string
        "400":
          description: "Atributos de usuario, correo contraseña son obligatorios"
          content:
            '*/*':
              schema:
                type: object
  /publico/cambiarContraseña:
    post:
      tags:
      - Cambio de clave
      summary: Cambiar contrasena
      description: "Mediante el ingreso del codigo recibido por email, permiteestabloecer\
        \ una nueva contrasena"
      operationId: cambiarClave
      parameters:
      - in: query
        description: Datos para cambiar la clave
        required: true
        example: "usuario:'Juan',clave:'1234',codigo:'435678'"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CambioClaveDTO"
        required: true
      responses:
        "200":
          description: Clave cambiada
          content:
            application/json:
              schema:
                type: string
        "404":
          description: Usuario no encontrado
          content:
            '*/*':
              schema:
                type: object
        "403":
          description: Codigo incorrecto
          content:
            '*/*':
              schema:
                type: object
  /admin/get-all:
    get:
      tags:
      - Obtener Usuarios
      summary: Obtener usaurios
      description: Obtiene una pagina de los usuarios que hay en el sistem
      operationId: obtenerUsuarios
      parameters:
      - name: pagina
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - in: query
        description: Numero de pagina
        required: true
        example: 2
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
      security:
      - bearerAuth: []
components:
  schemas:
    DatosUsuario:
      required:
      - clave
      - correo
      - usuario
      type: object
      properties:
        usuario:
          type: string
        correo:
          type: string
        clave:
          type: string
    CambioClaveDTO:
      required:
      - clave
      - codigo
      - usuario
      type: object
      properties:
        usuario:
          type: string
        clave:
          type: string
        codigo:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      in: header
      scheme: bearer
      bearerFormat: JWT